decomp: revert Category C regressions from stash recovery
25 regressions from inadvertent refactoring reverted:
- Full checkout (11 files): FlowPtr, FlowTimer, FlowAnimate, UIListState,
  UIListDir, MemTrack, Joypad_Xinput, DanceRemixer, SuperEasyRemixer,
  MetagameRank, HamUI
- Flow.cpp: restore Enter() (ProxyFile check) + re-add deleted Exit()
- UI.cpp: restore FakeKeyboardAction constructor arg order + Handle() call
- UITrigger.cpp: revert std::fabs -> fabsf (keep d>> fix)
- FlowTrigger.cpp: restore operator>>, if/if structure, indentation
- Sequence.cpp: restore RandomIntervalGroupSeqInst::Poll() body (keep
  unsigned int ComputeNextTime + Stop FOREACH)
- UIScreen.cpp: restore if/else in Draw, DataDir/ObjectDir in ReloadStrings
  (keep #ifdef HX_NATIVE null panel check)
- UIFontImporter.cpp: restore GetGenericFont if/else, FindTextForFont font cmp
- InlineHelp.cpp: restore DrawShowing (DegreesToRadians/mTemplateLabel)
- UIListWidget.h + UIListSlot.h: restore inline Poll(){}/~dtor{} bodies
- Remove compensating stubs from UIListWidget.cpp, UIListSlot.cpp, HamNavList.cpp

Result: 29 regressions -> 5 (all Category B new decomp work, expected)

+++ src/system/ui/UITrigger.cpp
@@ -70,7 +70,7 @@ void UITrigger::Trigger() {
                     f4 = curAnim.mScale;
                     if (!f4)
                         f4 = 1.0f;
-                    f4 = std::fabs(curAnim.mStart - curAnim.mEnd) / f4;
+                    f4 = fabsf(curAnim.mStart - curAnim.mEnd) / f4;
                 }
             } else {
                 f4 = std::fabs(curAnim.mAnim->StartFrame() - curAnim.mAnim->EndFrame());
